requirements.txt
flask==2.0.1
Werkzeug==2.0.1
PyPDF2==3.0.1
Pillow==9.5.0
Markdown
google-generativeai==0.3.0
pytesseract==0.3.10
gunicorn==20.1.0
edge_tts==6.1.12
pydub==0.25.1